Create ESM and Commonjs flavor of config and env files
