+++ packages/sharing-smart-contract/config/config.cjs
@@ -1,0 +1,8 @@
+// Hardhat Ignition does not support ESM modules, so we use CommonJS syntax.
+// TODO refactor this to use ESM syntax when Hardhat Ignition supports it.
+
+module.exports = {
+    POCO_ADDRESS: '0x3eca1B216A7DF1C7689aEb259fFB83ADFB894E7f',
+    DATASET_REGISTRY_ADDRESS: '0x799DAa22654128d0C64d5b79eac9283008158730',
+    APP_REGISTRY_ADDRESS: '0xB1C52075b276f87b1834919167312221d50c9D16',
+};
