+++ packages/sharing-smart-contract/config/config.js
@@ -1,4 +1,5 @@
-export const SMART_CONTRACT_ADDRESS_FILE = '.smart-contract-address';
-export const POCO_ADDRESS = '0x3eca1B216A7DF1C7689aEb259fFB83ADFB894E7f';
-export const DATASET_REGISTRY_ADDRESS = '0x799DAa22654128d0C64d5b79eac9283008158730';
-export const APP_REGISTRY_ADDRESS = '0xB1C52075b276f87b1834919167312221d50c9D16';
+import config from './config.cjs';
+
+export const POCO_ADDRESS = config.POCO_ADDRESS;
+export const DATASET_REGISTRY_ADDRESS = config.DATASET_REGISTRY_ADDRESS;
+export const APP_REGISTRY_ADDRESS = config.APP_REGISTRY_ADDRESS;
